+++ receiver/k8sclusterreceiver/internal/collection/collector.go
@@ -21,6 +21,7 @@ import (
 	agentmetricspb "github.com/census-instrumentation/opencensus-proto/gen-go/agent/metrics/v1"
 	quotav1 "github.com/openshift/api/quota/v1"
 	"go.opentelemetry.io/collector/model/pdata"
+	"go.opentelemetry.io/collector/service/featuregate"
 	"go.uber.org/zap"
 	appsv1 "k8s.io/api/apps/v1"
 	autoscalingv2beta2 "k8s.io/api/autoscaling/v2beta2"
@@ -64,6 +65,8 @@ const (
 	k8sKindReplicationController = "ReplicationController"
 	k8sKindReplicaSet            = "ReplicaSet"
 	k8sStatefulSet               = "StatefulSet"
+
+	reportCPUMetetricsAsDouble = "receiver.k8sclusterreceiver.reportcpumetricsasdouble"
 )
 
 // DataCollector wraps around a metricsStore and a metadaStore exposing
@@ -78,8 +81,43 @@ type DataCollector struct {
 	allocatableTypesToReport []string
 }
 
+var reportCPUMetricsAsDoubleFeatureGate = featuregate.Gate{
+	// --feature-gates=receiver.k8sclusterreceiver.reportcpumetricsasdouble
+	ID:      reportCPUMetetricsAsDouble,
+	Enabled: false,
+	Description: "The k8s container and node cpu metrics being reported by the k8sclusterreceiver are transitioning " +
+		"from being reported as integer millicpu units to being reported as double cpu units to adhere to " +
+		"opentelemetry cpu metric specifications. Please update any monitoring this might affect, the change will " +
+		"cause cpu metrics to be double instead of integer values as well as metric values will be scaled down by " +
+		"1000x. You can control whether the k8sclusterreceiver reports container and node cpu metrics in double cpu " +
+		"units instead of integer millicpu units with the feature gate listed below. Note, feature gate identifiers " +
+		"prefixed with - will disable the gate and prefixing with + or with no prefix will enable the gate.\n" +
+		"Feature Gate Name:\n" +
+		"\treceiver.k8sclusterreceiver.reportcpumetricsasdouble\n" +
+		"Start otelcol with feature gate enabled:\n" +
+		"\totelcol {other_arguments} --feature-gates=receiver.k8sclusterreceiver.reportcpumetricsasdouble\n" +
+		"Start otelcol with feature gate disabled:\n" +
+		"\totelcol {other_arguments} --feature-gates=-receiver.k8sclusterreceiver.reportcpumetricsasdouble\n" +
+		"See more details:\n" +
+		"\thttps://github.com/open-telemetry/opentelemetry-collector-contrib/issues/8115\n" +
+		"\thttps://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/semantic_conventions/system-metrics.md#systemcpu---processor-metrics\n",
+}
+
+func init() {
+	featuregate.Register(reportCPUMetricsAsDoubleFeatureGate)
+}
+
 // NewDataCollector returns a DataCollector.
 func NewDataCollector(logger *zap.Logger, nodeConditionsToReport, allocatableTypesToReport []string) *DataCollector {
+	logger.Info("WARNING - Breaking Change: " + reportCPUMetricsAsDoubleFeatureGate.Description)
+	if featuregate.IsEnabled(reportCPUMetetricsAsDouble) {
+		logger.Info("The receiver.k8sclusterreceiver.reportcpumetricsasdouble feature gate is enabled. This " +
+			"otel collector will report double cpu units, which is good for future support!")
+	} else {
+		logger.Info("The receiver.k8sclusterreceiver.reportcpumetricsasdouble feature gate is disabled. This " +
+			"otel collector will report integer cpu units, be aware this will not be supported in the future.")
+	}
+
 	return &DataCollector{
 		logger: logger,
 		metricsStore: &metricsStore{